+++ test/libsolidity/semanticTests/state/block_difficulty_post_paris.sol
@@ -1,0 +1,12 @@
+contract C {
+    function f() public returns (uint) {
+        return block.difficulty;
+    }
+}
+// ====
+// compileToEwasm: also
+// EVMVersion: >=paris
+// ----
+// f() -> 0xa86c2e601b6c44eb4848f7d23d9df3113fbcac42041c49cbed5000cb4f118777
+// f() -> 0xa86c2e601b6c44eb4848f7d23d9df3113fbcac42041c49cbed5000cb4f118777
+// f() -> 0xa86c2e601b6c44eb4848f7d23d9df3113fbcac42041c49cbed5000cb4f118777